requirements.txt
python-telegram-bot
flask
gunicorn
snownlp
ytmusicapi